+++ src/app/page.tsx
@@ -3,15 +3,18 @@
 import Link from "next/link";
 import UserInfo from "../components/UserInfo";
 import About from "./about/page";
+import getUser from "@/lib/getUser";
 
 const Home = async () => {
+  const user = await getUser();
+
   return (
     <main>
       <h1>블로그</h1>
 
       <About />
       <Link href="/posts/create">post</Link>
-      <UserInfo />
+      <UserInfo user={user} />
     </main>
   );
 };
